pyproject.toml
[project]
name = "realtime-voice-assistant"
version = "1.0.0"
description = "Low-latency real-time voice conversation system"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
requires-python = ">=3.9"
dependencies = [
    # Core Server
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",
    "websockets>=12.0",

    # VAD (Voice Activity Detection)
    "silero-vad>=4.0.1",
    "torch>=2.0.0",
    "torchaudio>=2.0.0",

    # STT (Speech-to-Text)
    "mlx-whisper>=0.4.0",  # Apple Silicon optimized
    "numpy>=1.24.3",

    # LLM Inference - Ollama
    "ollama>=0.1.0",
    "httpx>=0.25.0",

    # TTS (Text-to-Speech)
    "piper-tts>=1.2.0",

    # Audio Processing
    "soundfile>=0.12.1",
    "pyaudio>=0.2.13",
    "webrtcvad>=2.0.10",

    # Utilities
    "python-dotenv>=1.0.0",
    "loguru>=0.7.2",
    "aiofiles>=23.2.1",
    "pydantic>=2.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

gpu = [
    "vllm>=0.3.0",
]

# tts-high-quality = [
#     "TTS>=0.20.0",  # Coqui TTS - conflicts with numpy requirements
#     "phonemizer>=3.2.1",
#     "scipy>=1.10.0",
#     "librosa>=0.10.0",
# ]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["server"]

[tool.uv]
dev-dependencies = []

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.black]
line-length = 100
target-version = ['py39']
